requirements.txt
BeautifulSoup4
requests
PDFMiner.six
click
pandas
